Use fs-extra's writeJson over own utils.writeFile
We already depend on fs-extra, and in the places where we used
utils.writeFile do have any special requirements about the formatting,
so switch those to fs-extra's writeJson.

Fold what remains of utils.js into update-bcd.js.
