+++ package.json
@@ -58,7 +58,6 @@
     "klaw": "3.0.0",
     "@mdn/browser-compat-data": "2.0.0",
     "mocha": "8.1.3",
-    "mock-fs": "4.13.0",
     "nodemon": "2.0.4",
     "nyc": "15.1.0",
     "open-cli": "6.0.1",
