+++ update-bcd.js
@@ -6,9 +6,31 @@ const klaw = require('klaw');
 const path = require('path');
 const uaParser = require('ua-parser-js');
 
-const {isEquivalent} = require('./utils');
-
 const overrides = require('./overrides').filter(Array.isArray);
+
+const isEquivalent = (a, b) => {
+  // Create arrays of property names
+  const aProps = Object.getOwnPropertyNames(a);
+  const bProps = Object.getOwnPropertyNames(b);
+
+  // If number of properties is different,
+  // objects are not equivalent
+  if (aProps.length != bProps.length) {
+    return false;
+  }
+
+  for (const propName of aProps) {
+    // If values of same property are not equal,
+    // objects are not equivalent
+    if (a[propName] !== b[propName]) {
+      return false;
+    }
+  }
+
+  // If we made it this far, objects
+  // are considered equivalent
+  return true;
+};
 
 const findEntry = (bcd, path) => {
   if (!path) {
@@ -393,6 +415,7 @@ if (require.main === module) {
   });
 } else {
   module.exports = {
+    isEquivalent,
     findEntry,
     getBrowserAndVersion,
     getSupportMap,
